Enable running view in separate process
Running the app from within PyCharm was problematic, but very much
desired. It was problematic because the PyCharm console has a bunch
of quirks that makes it not very much like a genuine terminal
emulator. The most annoying is that it handles certain keyboard
shortcuts so the app is not able to use them.

Now, by running the app with the appropriate flag, a dedicated
console window will appear and the app's view will appear there.